Revert "pf: remove pfsense-local pf_route() changes"
This reverts commit 353714427aebc9a04700a62fc047f5aa47d589eb.

Undo the removal of this local patch. It's not (or not just) intended to deal
with transparent proxies, but also changes route-to/reply-to so they do not send
to the specified gateway for subnet-local traffic.

That is, traffic from 192.0.2.2 on rule "pass in on $foo reply-to ($foo
192.0.2.3)" would normally see its replies sent to 192.0.2.3, but with this
patch we reply to 192.0.2.2 instead.
Arguably that sort of behaviour should be implemented in the pf ruleset, not as
an exception to route-to/reply-to's behaviour, but that's a non-trivail change
to the pfSense PHP code.